tests/CharacterData-insertData.html

<!DOCTYPE html>
<title>CharacterData.insertData</title>
<link rel="author" title="annevk" href="mailto:annevk@opera.com">
<link rel="help" href="../web-dom-core#dom-CharacterData-insertData">
<script src="testharness.js"></script>
<div id="log"></div>
<script>
function testNode(node) {
  test(function() {
    assert_throws("INDEX_SIZE_ERR", function() { node.insertData(5, "x") })
    assert_throws("INDEX_SIZE_ERR", function() { node.insertData(5, "") })
    node.insertData(2, "X")
    assert_equals(node.data, "teXst")
    node.data = "test"
    node.insertData(4, "ing")
    assert_equals(node.data, "testing")
  })
}
test(function() {
  testNode(document.createTextNode("test"))
  testNode(document.createComment("test"))
})
</script>
